% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{backward_correction}
\alias{backward_correction}
\title{Backward Selection for Shift Correction}
\usage{
backward_correction(
  tree,
  Y,
  alpha,
  sv_mean,
  sv_var,
  criterion = "BIC",
  original_model = NULL,
  measurement_error = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tree}{Phylogenetic tree.}

\item{Y}{Trait values.}

\item{alpha}{Selection strength.}

\item{sv_mean}{Mean shift branches.}

\item{sv_var}{Variance shift branches.}

\item{criterion}{\code{"BIC"}, \code{"mBIC"}, or \code{"pBIC"}.}

\item{original_model}{Optional pre-fitted model.}

\item{measurement_error}{Logical.}

\item{verbose}{Logical.}
}
\value{
Refined model.
}
\description{
Backward Selection for Shift Correction
}
