% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGM-package.R
\docType{package}
\name{EGM-package}
\alias{EGM}
\alias{EGM-package}
\title{EGM: Evaluating Cardiac Electrophysiology Signals}
\description{
A system for importing electrophysiological signal into R, based on the Waveform Database(WFDB) software package. This package provides a wrapper for utilizing WFDB functions for reading and writing signal data in R, as well as functions for visualization and analysis. This package introduces a stable and compatible class for working with signal data, and supports reading in of specific, electrophysiogical file formats, including intracardiac electrograms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shah-in-boots.github.io/EGM/}
  \item Report bugs at \url{https://github.com/shah-in-boots/EGM/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
