% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{generate_design_matrix}
\alias{generate_design_matrix}
\title{Generate Design Matrix}
\usage{
generate_design_matrix(tree, type = "simpX", alpha = 0)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{type}{A character string: \code{"simpX"} or \code{"orgX"}.}

\item{alpha}{Selection strength (only for \code{type = "orgX"}).}
}
\value{
A design matrix X.
}
\description{
Constructs a design matrix for a given phylogenetic tree.
}
