% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{OU.vcv}
\alias{OU.vcv}
\title{OU.vcv}
\usage{
OU.vcv(tree, alpha)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{alpha}{selection strength}
}
\value{
Covariance matrix V.
}
\description{
Generate covariance matrix for OU process.
}
