% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{plot.ShiftModel}
\alias{plot.ShiftModel}
\title{Plot Method for ShiftModel Objects}
\usage{
\method{plot}{ShiftModel}(x, title = "", ...)
}
\arguments{
\item{x}{A ShiftModel object.}

\item{title}{Plot title.}

\item{...}{Additional arguments passed to plot.phylo.}
}
\value{
No return value (side effect: plot).
}
\description{
Plots a phylogenetic tree with detected shifts highlighted.
}
