% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{print.summary.ShiftModel}
\alias{print.summary.ShiftModel}
\title{Print Method for Summary of ShiftModel}
\usage{
\method{print}{summary.ShiftModel}(x, ...)
}
\arguments{
\item{x}{A summary.ShiftModel object.}

\item{...}{Unused.}
}
\value{
No return value (side effect: console output).
}
\description{
Print Method for Summary of ShiftModel
}
