% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-structures.R
\name{signal_table}
\alias{signal_table}
\alias{is_signal_table}
\title{Signal tables}
\usage{
signal_table(...)

is_signal_table(x)
}
\arguments{
\item{...}{A \code{list} of equal lengths}

\item{x}{\code{data.frame} A data frame}
}
\description{
\code{signal_table()} modifies the \code{data.table} class to work with electrical
signal data. The input should be a data set of equal number of rows. It will
add a column of index positions called \code{sample} if it does not already exist.
}
