% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{soft_thresholding}
\alias{soft_thresholding}
\title{Soft Thresholding}
\usage{
soft_thresholding(z, lambda)
}
\arguments{
\item{z}{Numeric value.}

\item{lambda}{Threshold level.}
}
\value{
Thresholded value.
}
\description{
Soft Thresholding
}
