% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{summary.ShiftModel}
\alias{summary.ShiftModel}
\title{Summary of a ShiVa Shift Model}
\usage{
\method{summary}{ShiftModel}(object, ...)
}
\arguments{
\item{object}{A ShiftModel object.}

\item{...}{Unused.}
}
\value{
A summary.ShiftModel object.
}
\description{
Summary of detected shifts and fitted parameters.
}
