% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{summary_shifts}
\alias{summary_shifts}
\title{Summary of Shifts from a ShiVa Model Result}
\usage{
summary_shifts(result)
}
\arguments{
\item{result}{A list returned by the \code{ShiVa()} function, containing at least \code{best_model} and \code{tree}.}
}
\description{
Print a concise summary of the best model selected by \code{ShiVa()}, including shift locations (edges) and magnitudes, as well as estimated parameters.
}
