% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{update_step_gamma}
\alias{update_step_gamma}
\title{Update Step for Gamma}
\usage{
update_step_gamma(gamma_k, X_k, Sigma, r, lambda2, t, penalty, V, q_k)
}
\arguments{
\item{gamma_k}{Current gamma value.}

\item{X_k}{Column of design matrix.}

\item{Sigma}{Current covariance matrix.}

\item{r}{Residual vector.}

\item{lambda2}{Penalty parameter.}

\item{t}{Step size.}

\item{penalty}{\code{"L1"} or \code{"None"}.}

\item{V}{Baseline covariance matrix.}

\item{q_k}{Element of design vector q.}
}
\value{
List with updated gamma_k and Sigma.
}
\description{
Update Step for Gamma
}
