% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfdb-utils.R
\name{wfdb_paths}
\alias{wfdb_paths}
\alias{find_wfdb_software}
\alias{set_wfdb_path}
\alias{find_wfdb_command}
\title{WFDB path utilities}
\usage{
find_wfdb_software()

set_wfdb_path(.path)

find_wfdb_command(.app, .path = getOption("wfdb_path"))
}
\arguments{
\item{.path}{A \code{character} string that describes the path to the WFDB binary
directory}

\item{.app}{The name of WFDB software command or application as a \code{character}}
}
\description{
These functions are used to help find and locate commands from the
installation of WFDB. They are helpful in setting and getting path options
and specific WFDB commands. They are primarily internal helper functions, but
are documented for troubleshooting purposes.
}
