% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{backward_correction}
\alias{backward_correction}
\title{backward_correction}
\usage{
backward_correction(
  tree,
  Y,
  alpha,
  sv_mean,
  sv_var,
  criterion = "BIC",
  original_model = NULL
)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{Y}{a vector of trait values}

\item{alpha}{the parameter for selection strength}

\item{sv_mean}{shifts in optimal value}

\item{sv_var}{shifts in variance}

\item{criterion}{measurement criterion}

\item{original_model}{original OU model; optional, default NULL}
}
\value{
\item{OModel}{Fitted OU model}
}
\description{
This function conduct backward selection on given combination of shifts in optimal value and variance
}
