% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{fit_OU_mean_var}
\alias{fit_OU_mean_var}
\title{fit_OU_mean_var}
\usage{
fit_OU_mean_var(
  tree,
  Y,
  alpha,
  sv_mean,
  sv_var,
  max.steps = 1000,
  t = 0.01,
  thres = 0.01
)
}
\arguments{
\item{tree}{a phylogenetic tree}

\item{Y}{a vector of trait values}

\item{alpha}{the parameter for selection strength}

\item{sv_mean}{shifts in optimal value}

\item{sv_var}{shifts in variance}

\item{max.steps}{maximum number of steps, default 1000}

\item{t}{step size, default 0.01}

\item{thres}{threshold for stopping optimization}
}
\value{
\item{beta}{Shift sizes for optimal value}
\item{gamma}{Shift sizes for variance}
\item{sigma2}{Original variance without shifts}
\item{b0}{The intercept term}
\item{loglik}{log-likelihood of the fitted model}
\item{BIC}{BIC of the fitted model}
\item{fitted.values}{fitted trait values}
\item{Sigma}{estimated covariance matrix}
}
\description{
This function provide estimation of shift sizes in optimal value and shifts in variance given estimated shifts
}
