% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{generate_design_matrix}
\alias{generate_design_matrix}
\title{generate_design_matrix}
\usage{
generate_design_matrix(tree, type = "simpX", alpha)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{type}{type of design matrix}

\item{alpha}{parameter alpha}
}
\value{
\item{X}{desigh matrix}
}
\description{
generate design matrix
}
