% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{generate_design_matrix}
\alias{generate_design_matrix}
\title{generate_design_matrix}
\usage{
generate_design_matrix(tree)
}
\arguments{
\item{tree}{phylogenetic tree}
}
\value{
\item{X}{design matrix}
}
\description{
generate design matrix
}
