% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{get_mean_var_shifts}
\alias{get_mean_var_shifts}
\title{get_mean_var_shifts}
\usage{
get_mean_var_shifts(
  Y,
  tree,
  alpha,
  lambda1,
  lambda2,
  max.steps = 1000,
  t = 0.01,
  penalty = "L1",
  thres = 0.001,
  sigma2 = NULL
)
}
\arguments{
\item{Y}{a vector of trait values}

\item{tree}{a phylogenetic tree}

\item{alpha}{the parameter for selection strength}

\item{lambda1}{The degree of shrinkage for beta (shifts in optimal value)}

\item{lambda2}{The degree of shrinkage for gamma (shifts in variance)}

\item{max.steps}{maximum number of steps, default 1000}

\item{t}{step size, default 0.01}

\item{penalty}{"L1" or "None"}

\item{thres}{threshold for stopping optimization}

\item{sigma2}{the original sigma2, default is unknown.}
}
\value{
\item{beta}{Shifts in optimal value}
\item{gamma}{Shifs in variance}
\item{sigma2}{Original variance without shifts}
\item{b0}{The intercept term}
}
\description{
This function provide estimation of shifts in optimal value and shifts in variance given fixed lambda1 and lambda2
}
