% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{get_mean_var_shifts_model_selection}
\alias{get_mean_var_shifts_model_selection}
\title{get_mean_var_shifts_model_selection}
\usage{
get_mean_var_shifts_model_selection(
  Y,
  tree,
  alpha,
  lambda1_list = NULL,
  lambda2_list,
  criterion = "BIC",
  max.steps = 100,
  nfolds = 8,
  top_k = 10
)
}
\arguments{
\item{Y}{a vector of trait values}

\item{tree}{a phylogenetic tree}

\item{alpha}{the parameter for selection strength}

\item{lambda1_list}{a list of degree of shrinkage parameters for shifts in optimal value}

\item{lambda2_list}{a list of degree of shrinkage parameters for shifts in variance}

\item{criterion}{the criterion for model selection, default BIC}

\item{max.steps}{maximum number of steps, default 100}

\item{nfolds}{number of folds for cross validation for tunning lambda1, default 8}

\item{top_k}{conduct backward correction on k original models with the best criterion scores, default 10}
}
\value{
\item{best_model}{Fitted OU model}
\item{score_summary}{score summary table}
}
\description{
This function provide estimation of shift sizes in optimal value and shifts in variance (including model selection)
}
\examples{
require(ape)
require(MASS)
require(PIGShift)
alpha = 1
sigma2_0 = alpha*2
true_shifts_var = c(8) #shift configuration in variance
size_var = 15 #shift size in variance
true_shifts_mean = 0 #shift configuration in mean
size_mean = 0 #shift size in mean
set.seed(345)
# generate a tree
tree = rcoal(20)
tree$edge.length = tree$edge.length/max(node.depth.edgelength(tree))
# generate design matrix
X = generate_design_matrix(tree)
V = OU.vcv(tree,alpha)
tb = node.depth.edgelength(tree)[tree$edge[,1]]
q = exp(-2*alpha*max(node.depth.edgelength(tree)))/(2*alpha)*(1-exp(2*alpha*tb))
# simulate shifts in variance
gamma = rep(0,ncol(X))
gamma[true_shifts_var]  = size_var
Sigma = V*sigma2_0 +  X\%*\%diag(gamma)\%*\%t(X)*V+ X\%*\%diag(gamma*q)\%*\%t(X)
# simulate shifts in mean
beta = rep(0,ncol(X))
beta[true_shifts_mean] = size_mean
# generate simulated data
eps = mvrnorm(n = 1, mu = rep(0,nrow(X)), Sigma = Sigma)
Y = X\%*\%beta+eps
result = get_mean_var_shifts_model_selection(Y,tree,alpha,NULL,exp(1:10-6),nfolds=3)
}
