% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{OU.vcv}
\alias{OU.vcv}
\title{OU.vcv}
\usage{
OU.vcv(tree, alpha)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{alpha}{selective force}
}
\value{
\item{V}{covariance matrix}
}
\description{
generate covariance matrix for OU process
}
