% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{soft_thresholding}
\alias{soft_thresholding}
\title{soft_thresholding}
\usage{
soft_thresholding(z, lambda)
}
\arguments{
\item{z}{input value}

\item{lambda}{degree of shrinkage}
}
\value{
\item{new_value}{The result of soft thresholding function}
}
\description{
Soft thresholding function
}
