% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiVa.R
\name{update_step_gamma}
\alias{update_step_gamma}
\title{update_step_gamma}
\usage{
update_step_gamma(gamma_k, X_k, Sigma, r, lambda2, t, penalty, V, q_k)
}
\arguments{
\item{gamma_k}{The gamma value to be updated}

\item{X_k}{The kth column of the design matrix}

\item{Sigma}{The covariance matrix}

\item{r}{The residual vector (Y-X*beta)}

\item{lambda2}{The degree of shrinkage for gamma}

\item{t}{step size}

\item{penalty}{"L1" or "None"}

\item{V}{The original covariance matrix when sigma^2 = 1}

\item{q_k}{The kth element of designed vector q}
}
\value{
\item{gamma_k}{The updated gamma value}
\item{Sigma}{The updated covariance matrix}
}
\description{
This function calculate the updated gamma for each step
}
